user and shelf activities: make it fast: fetch all items at once
when having fetched activities documents, find maximum since and minimum
until, to only have one items query instead of one per activity

another implementation would be to fetch activities items since the time
the user went fediversable until now
